feat [backend/src/main/java] added classes Post, PostDTO, PostService, PostFacade and interface PostRepo
