--- backend/docker-compose.yml
@@ -1,25 +1,0 @@
-
-
-services:
-  db:
-    container_name: myDatabase
-    image: postgres
-    environment:
-      POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: postgres
-      POSTGRES_DB: user_db
-
-    volumes:
-      - db:/var/lib/postgresql/data
-    ports:
-      - "5432:5432"
-    networks:
-      - db
-    restart: unless-stopped
-
-networks:
-  db:
-    driver: bridge
-
-volumes:
-  db:
