--- backend/src/main/java/com/gleb/SocialNetworkApplication.java
@@ -1,19 +1,0 @@
-package com.gleb;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.data.r2dbc.repository.config.EnableR2dbcRepositories;
-import org.springframework.scheduling.annotation.EnableAsync;
-@SpringBootApplication
-@EnableCaching
-@EnableAsync
-@EnableR2dbcRepositories
-public class SocialNetworkApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SocialNetworkApplication.class, args);
-    }
-
-}
