+++ backend/src/main/java/com/gleb/service/UserService.java
@@ -10,7 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Mono;
 
 import java.util.Collections;
-import java.util.Optional;
 
 @Service
 @Transactional(readOnly = true)
