+++ backend/src/main/resources/application.yml
@@ -1,22 +1,6 @@
-server:
-  port: 8080
-
 spring:
-  datasource:
-    url: r2dbc:postgresql://localhost:5432/web-app
+  r2dbc:
+    url: r2dbc:postgresql://localhost:5432/userDb
     username: postgres
     password: password
-
-
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-        format_sql: true
-    show-sql: true
-
-  main:
-    web-application-type: servlet
-
+    name: userDb
