--- src/main/java/com/gleb/data/User/RoleName.java
@@ -1,5 +1,0 @@
-package com.gleb.data.User;
-
-public enum RoleName {
-    ROLE_USER, ROLE_MODERATOR, ROLE_ADMIN
-}
