--- src/main/java/com/gleb/dto/UserRegisterDTO.java
@@ -1,15 +1,0 @@
-package com.gleb.dto;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class UserRegisterDTO {
-    private String username;
-    private String password;
-    private String email;
-    private String firstName;
-    private String surname;
-    private LocalDate birthDate;
-}
