--- src/main/java/com/gleb/exceptions/UserAlreadyExistsException.java
@@ -1,9 +1,0 @@
-package com.gleb.exceptions;
-
-
-    public class UserAlreadyExistsException extends RuntimeException {
-        public UserAlreadyExistsException(String message) {
-            super(message);
-        }
-    }
-
