--- src/main/resources/application.yml
@@ -1,6 +1,0 @@
-spring:
-  r2dbc:
-    url: r2dbc:postgresql://localhost:5432/userDb
-    username: postgres
-    password: password
-    name: userDb
